+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v2/create-table-action.ts
@@ -5,7 +5,10 @@ import OverlayRegistry from 'sap/ui/dt/OverlayRegistry';
 import ManagedObject from 'sap/ui/base/ManagedObject';
 import UI5Element from 'sap/ui/core/Element';
 
-import { QuickActionContext, NestedQuickActionDefinition } from '../../../cpe/quick-actions/quick-action-definition';
+import {
+    QuickActionContext,
+    NestedQuickActionDefinition
+} from '../../../cpe/quick-actions/quick-action-definition';
 import { getControlById, isA } from '../../../utils/core';
 import { DialogFactory, DialogNames } from '../../dialog-factory';
 import { TableQuickActionDefinitionBase } from '../table-quick-action-base';
@@ -71,11 +74,18 @@ export class AddTableActionQuickAction extends TableQuickActionDefinitionBase im
         // open dialogBox to add, and content is selected ByDefault
         if (headerToolbar) {
             const overlay = OverlayRegistry.getOverlay(headerToolbar as UI5Element) || [];
-            await DialogFactory.createDialog(overlay, this.context.rta, DialogNames.ADD_FRAGMENT, undefined, {
-                aggregation: 'content',
-                title: 'QUICK_ACTION_ADD_CUSTOM_TABLE_ACTION',
-                defaultAggregationArrayIndex: 1
-            });
+            await DialogFactory.createDialog(
+                overlay,
+                this.context.rta,
+                DialogNames.ADD_FRAGMENT,
+                undefined,
+                {
+                    aggregation: 'content',
+                    title: 'QUICK_ACTION_ADD_CUSTOM_TABLE_ACTION',
+                    defaultAggregationArrayIndex: 1
+                },
+                { actionName: this.type, telemetryEventIdentifier: this.getTelemetryIdentifier() }
+            );
         }
         return [];
     }