+++ packages/preview-middleware-client/test/__mock__/sap/ui/rta/util/isReuseComponent.ts
@@ -1,0 +1,3 @@
+export default {
+    isReuseComponent: jest.fn()
+}