+++ packages/preview-middleware-client/test/unit/adp/quick-actions/fe-v4.test.ts
@@ -1,7 +1,10 @@
-import RuntimeAuthoring, { RTAOptions } from 'sap/ui/rta/RuntimeAuthoring';
+import RuntimeAuthoring, { FlexSettings, RTAOptions } from 'sap/ui/rta/RuntimeAuthoring';
 import FlexBox from 'sap/m/FlexBox';
 import RuntimeAuthoringMock from 'mock/sap/ui/rta/RuntimeAuthoring';
 import { attachBeforeClose } from 'mock/sap/ui/core/Fragment';
+import ODataModelV4 from 'sap/ui/model/odata/v4/ODataModel';
+import type AppComponentV4 from 'sap/fe/core/AppComponent';
+
 import type { ChangeService } from '../../../../src/cpe/changes/service';
 const mockChangeService = {
     syncOutlineChanges: jest.fn()
@@ -46,6 +49,13 @@ import { TableQuickActionDefinitionBase } from '../../../../src/adp/quick-action
 import * as QCUtils from '../../../../src/cpe/quick-actions/utils';
 import ManagedObject from 'sap/ui/base/ManagedObject';
 import * as versionUtils from 'open/ux/preview/client/utils/version';
+import * as utils from 'open/ux/preview/client/utils/fe-v4';
+
+let telemetryEventIdentifier: string;
+const mockTelemetryEventIdentifier = () => {
+    telemetryEventIdentifier = new Date().toISOString();
+    jest.spyOn(Date.prototype, 'toISOString').mockReturnValue(telemetryEventIdentifier);
+};
 
 describe('FE V4 quick actions', () => {
     let sendActionMock: jest.Mock;
@@ -137,7 +147,7 @@ describe('FE V4 quick actions', () => {
                 });
                 jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                     if (id === 'component-id') {
-                        return component;
+                        return component as unknown as ComponentMock;
                     }
                 });
                 container.getCurrentPage.mockImplementation(() => {
@@ -266,7 +276,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -285,14 +295,19 @@ describe('FE V4 quick actions', () => {
 
                 rtaMock = new RuntimeAuthoringMock({} as RTAOptions) as unknown as RuntimeAuthoring;
                 const registry = new FEV4QuickActionRegistry();
-                service = new QuickActionService(rtaMock, new OutlineService(rtaMock, mockChangeService), [registry], {
-                    onStackChange: jest.fn(),
-                    getConfigurationPropertyValue: jest
-                        .fn()
-                        .mockReturnValueOnce(false)
-                        .mockReturnValueOnce(undefined)
-                        .mockReturnValue(undefined)
-                } as any);
+                service = new QuickActionService(
+                    rtaMock,
+                    new OutlineService(rtaMock, mockChangeService),
+                    [registry],
+                    {
+                        onStackChange: jest.fn(),
+                        getConfigurationPropertyValue: jest
+                            .fn()
+                            .mockReturnValueOnce(false)
+                            .mockReturnValueOnce(undefined)
+                            .mockReturnValue(undefined)
+                    } as any
+                );
             });
 
             test('initialize and execute action', async () => {
@@ -386,6 +401,7 @@ describe('FE V4 quick actions', () => {
         describe('add controller to the page', () => {
             test('initialize and execute action', async () => {
                 const pageView = new XMLView();
+                mockTelemetryEventIdentifier();
                 FlexUtils.getViewForControl.mockImplementation(() => {
                     return {
                         getId: () => 'MyView',
@@ -442,7 +458,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -498,7 +514,14 @@ describe('FE V4 quick actions', () => {
                     executeQuickAction({ id: 'listReport0-add-controller-to-page', kind: 'simple' })
                 );
 
-                expect(DialogFactory.createDialog).toHaveBeenCalledWith(mockOverlay, rtaMock, 'ControllerExtension');
+                expect(DialogFactory.createDialog).toHaveBeenCalledWith(
+                    mockOverlay,
+                    rtaMock,
+                    'ControllerExtension',
+                    undefined,
+                    {},
+                    { actionName: 'add-controller-to-page', telemetryEventIdentifier }
+                );
             });
         });
 
@@ -542,7 +565,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -690,7 +713,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -822,7 +845,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -907,7 +930,8 @@ describe('FE V4 quick actions', () => {
                     {
                         aggregation: 'columns',
                         title: 'QUICK_ACTION_ADD_CUSTOM_TABLE_COLUMN'
-                    }
+                    },
+                    expect.objectContaining({ actionName: 'create-table-custom-column' })
                 );
             });
         });
@@ -979,7 +1003,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -1006,8 +1030,8 @@ describe('FE V4 quick actions', () => {
                             .mockReturnValueOnce(undefined)
                     } as any
                 );
-                await service.init(sendActionMock, subscribeMock);
 
+                await service.init(sendActionMock, subscribeMock);
                 await service.reloadQuickActions({
                     'sap.ui.mdc.Table': [
                         {
@@ -1176,7 +1200,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -1195,16 +1219,21 @@ describe('FE V4 quick actions', () => {
 
                 rtaMock = new RuntimeAuthoringMock({} as RTAOptions) as unknown as RuntimeAuthoring;
                 const registry = new FEV4QuickActionRegistry();
-                service = new QuickActionService(rtaMock, new OutlineService(rtaMock, mockChangeService), [registry], {
-                    onStackChange: jest.fn(),
-                    getConfigurationPropertyValue: jest
-                        .fn()
-                        .mockReturnValueOnce(undefined)
-                        .mockReturnValueOnce(undefined)
-                        .mockReturnValueOnce(true)
-                        .mockReturnValueOnce(undefined)
-                        .mockReturnValue(undefined)
-                } as any);
+                service = new QuickActionService(
+                    rtaMock,
+                    new OutlineService(rtaMock, mockChangeService),
+                    [registry],
+                    {
+                        onStackChange: jest.fn(),
+                        getConfigurationPropertyValue: jest
+                            .fn()
+                            .mockReturnValueOnce(undefined)
+                            .mockReturnValueOnce(undefined)
+                            .mockReturnValueOnce(true)
+                            .mockReturnValueOnce(undefined)
+                            .mockReturnValue(undefined)
+                    } as any
+                );
             });
 
             test('initialize and execute action', async () => {
@@ -1370,7 +1399,7 @@ describe('FE V4 quick actions', () => {
                         });
                         jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                             if (id === 'component-id') {
-                                return component;
+                                return component as unknown as ComponentMock;
                             }
                         });
                         container.getCurrentPage.mockImplementation(() => {
@@ -1539,6 +1568,7 @@ describe('FE V4 quick actions', () => {
                     }
                 ];
                 test.each(testCases)('initialize and execute action (%s)', async (testCase) => {
+                    mockTelemetryEventIdentifier();
                     const pageView = new XMLView();
                     FlexUtils.getViewForControl.mockImplementation(() => {
                         return {
@@ -1601,7 +1631,7 @@ describe('FE V4 quick actions', () => {
                             });
                             jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                                 if (id === 'component-id') {
-                                    return component;
+                                    return component as unknown as ComponentMock;
                                 }
                             });
                             container.getCurrentPage.mockImplementation(() => {
@@ -1674,7 +1704,8 @@ describe('FE V4 quick actions', () => {
                         {
                             aggregation: 'items',
                             title: 'QUICK_ACTION_OP_ADD_HEADER_FIELD'
-                        }
+                        },
+                        { actionName: 'op-add-header-field', telemetryEventIdentifier }
                     );
                 });
             });
@@ -1706,7 +1737,9 @@ describe('FE V4 quick actions', () => {
                         enable: true
                     }
                 ];
+
                 test.each(testCases)('initialize and execute action (%s)', async (testCase) => {
+                    mockTelemetryEventIdentifier();
                     const pageView = new XMLView();
                     const scrollIntoView = jest.fn();
                     jest.spyOn(TableQuickActionDefinitionBase.prototype as any, 'getInternalTable').mockImplementation(
@@ -1812,7 +1845,7 @@ describe('FE V4 quick actions', () => {
                             });
                             jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
                                 if (id === 'component-id') {
-                                    return component;
+                                    return component as unknown as ComponentMock;
                                 }
                             });
                             container.getCurrentPage.mockImplementation(() => {
@@ -1898,6 +1931,10 @@ describe('FE V4 quick actions', () => {
                         {
                             aggregation: 'columns',
                             title: 'QUICK_ACTION_ADD_CUSTOM_TABLE_COLUMN'
+                        },
+                        {
+                            actionName: 'create-table-custom-column',
+                            telemetryEventIdentifier
                         }
                     );
                 });
@@ -2078,7 +2115,7 @@ describe('FE V4 quick actions', () => {
                                 jest.spyOn(Component, 'getComponentById').mockImplementation(
                                     (id: string | undefined) => {
                                         if (id === 'component-id') {
-                                            return component;
+                                            return component as unknown as ComponentMock;
                                         }
                                     }
                                 );
@@ -2262,15 +2299,15 @@ describe('FE V4 quick actions', () => {
                             jest.spyOn(componentContainer, 'getComponent').mockImplementation(() => {
                                 return 'component-id';
                             });
-                            jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
+                            jest.spyOn(ComponentMock, 'getComponentById').mockImplementation((id: string) => {
                                 if (id === 'component-id') {
-                                    return component;
+                                    return component as unknown as ComponentMock;
                                 }
                             });
                             container.getCurrentPage.mockImplementation(() => {
                                 return componentContainer;
                             });
-                            component.getRootControl.mockImplementation(() => {
+                            jest.spyOn(component, 'getRootControl').mockImplementation(() => {
                                 return pageView;
                             });
                             jest.spyOn(ComponentMock, 'getOwnerComponentFor').mockImplementation(() => {
@@ -2430,6 +2467,434 @@ describe('FE V4 quick actions', () => {
                     }
                 });
             });
+        });
+    });
+
+    describe('Add subpage', () => {
+        const testCases: {
+            ui5version?: versionUtils.Ui5VersionInfo;
+            isNewPageUnavailable?: boolean;
+            isUnexpectedOwnerComponent?: boolean;
+            componentHasNoEntitySet?: boolean;
+            isListReport?: boolean;
+            isBetaFeatureDisabled?: boolean;
+            isContextPathDefined?: boolean;
+            isNoRouteFound?: boolean;
+            expect: {
+                toBeAvailable: boolean;
+                toBeEnabled?: boolean;
+                toThrow?: string;
+                tooltip?: string;
+            };
+        }[] = [
+            {
+                isListReport: true,
+                expect: {
+                    toBeAvailable: true,
+                    toBeEnabled: true
+                }
+            },
+            {
+                isListReport: true,
+                isNewPageUnavailable: true,
+                isContextPathDefined: true,
+                expect: {
+                    toBeAvailable: true,
+                    toBeEnabled: false,
+                    tooltip: `This option has been disabled because there are no subpages to add`
+                }
+            },
+            {
+                isListReport: true,
+                isNewPageUnavailable: true,
+                expect: {
+                    toBeAvailable: true,
+                    toBeEnabled: false,
+                    tooltip: `This option has been disabled because there are no subpages to add`
+                }
+            },
+            {
+                expect: {
+                    toBeAvailable: true,
+                    toBeEnabled: true
+                }
+            },
+            {
+                isNewPageUnavailable: true,
+                expect: {
+                    toBeAvailable: true,
+                    toBeEnabled: false,
+                    tooltip: `This option has been disabled because there are no subpages to add`
+                }
+            },
+            {
+                isNewPageUnavailable: true,
+                isContextPathDefined: true,
+                expect: {
+                    toBeAvailable: true,
+                    toBeEnabled: false,
+                    tooltip: `This option has been disabled because there are no subpages to add`
+                }
+            },
+            {
+                ui5version: {
+                    major: 1,
+                    minor: 133
+                },
+                expect: {
+                    toBeAvailable: false
+                }
+            },
+            {
+                isUnexpectedOwnerComponent: true,
+                expect: {
+                    toBeAvailable: false
+                }
+            },
+            {
+                componentHasNoEntitySet: true,
+                expect: {
+                    toBeAvailable: false
+                }
+            },
+            {
+                isNoRouteFound: true,
+                expect: {
+                    toBeAvailable: false
+                }
+            }
+        ];
+        afterEach(() => {
+            jest.restoreAllMocks();
+        });
+        test.each(testCases)('initialize and execute action (%s)', async (testCase) => {
+            mockTelemetryEventIdentifier();
+            jest.spyOn(versionUtils, 'getUi5Version').mockResolvedValue(
+                testCase.ui5version ?? { major: 1, minor: 135 }
+            );
+            jest.spyOn(FeatureService, 'isFeatureEnabled').mockReturnValue(!testCase.isBetaFeatureDisabled);
+
+            const pageView = new XMLView();
+            pageView.getParent.mockReturnValue({
+                getProperty: (propName: string) => {
+                    if (propName === 'entitySet') {
+                        return 'DummyEntitySet';
+                    } else {
+                        return undefined;
+                    }
+                }
+            });
+
+            const actionId = testCase.isListReport ? 'listReport0-add-new-subpage' : 'objectPage0-add-new-subpage';
+
+            jest.spyOn(ComponentMock, 'getOwnerComponentFor').mockImplementation(() => {
+                return {
+                    isA: (type: string) =>
+                        type ===
+                        (testCase.isUnexpectedOwnerComponent ? 'wrongType' : 'sap.fe.templates.ListReport.Component'),
+                    getEntitySet: jest
+                        .fn()
+                        .mockReturnValue(
+                            testCase.componentHasNoEntitySet ? undefined : testCase.isListReport ? 'Travel' : 'Booking'
+                        )
+                } as unknown as UIComponent;
+            });
+
+            sapCoreMock.byId.mockImplementation((id) => {
+                if (id == 'ObjectPage') {
+                    return {
+                        isA: (type: string) => type === 'sap.fe.templates.ObjectPage.Component',
+                        getId: () => id,
+                        getDomRef: () => ({ ref: 'OP' }),
+                        getParent: () => pageView
+                    };
+                }
+                if (id == 'ListReport') {
+                    return {
+                        isA: (type: string) => type === 'sap.fe.templates.ListReport.Component',
+                        getId: () => id,
+                        getDomRef: () => ({ ref: 'LR' }),
+                        getParent: () => pageView
+                    };
+                }
+                if (id == 'NavContainer') {
+                    const container = new NavContainer();
+                    const component = new ComponentMock();
+                    const view = new XMLView();
+                    pageView.getDomRef.mockImplementation(() => {
+                        return {
+                            contains: (domRef: { ref: string }) => domRef.ref === (testCase.isListReport ? 'LR' : 'OP')
+                        };
+                    });
+                    pageView.getViewName.mockImplementation(
+                        () =>
+                            `sap.fe.templates.${
+                                testCase.isListReport ? 'ListReport.ListReport' : 'ObjectPage.ObjectPage'
+                            }`
+                    );
+                    pageView.getViewData.mockImplementation(() => ({
+                        stableId: testCase.isListReport ? 'appId::TravelList' : 'appId::BookingObjectPage'
+                    }));
+
+                    jest.spyOn(view, 'getComponent').mockReturnValue('component-id');
+
+                    jest.spyOn(Component, 'getComponentById').mockImplementation((id: string | undefined) => {
+                        if (id === 'component-id') {
+                            return component;
+                        }
+                    });
+                    container.getCurrentPage.mockImplementation(() => {
+                        return view;
+                    });
+                    jest.spyOn(component, 'getRootControl').mockImplementation(() => {
+                        return pageView;
+                    });
+                    return container;
+                }
+            });
+
+            const rtaMock = new RuntimeAuthoringMock({} as RTAOptions) as unknown as RuntimeAuthoring;
+            const targets = {
+                TravelList: {
+                    'id': 'TravelList',
+                    'name': 'sap.fe.templates.ListReport',
+                    'options': {
+                        'settings': {
+                            'entitySet': 'Travel'
+                        }
+                    }
+                },
+                ...(testCase.isListReport && testCase.isNewPageUnavailable
+                    ? {
+                          TravelObjectPage: {
+                              'id': 'TravelObjectPage',
+                              'name': 'sap.fe.templates.ObjectPage',
+                              'options': {
+                                  'settings': testCase.isContextPathDefined
+                                      ? {
+                                            'contextPath': '/Travel'
+                                        }
+                                      : {
+                                            'entitySet': 'Travel'
+                                        }
+                              }
+                          }
+                      }
+                    : {}),
+                BookingObjectPage: {
+                    'id': 'BookingObjectPage',
+                    'name': 'sap.fe.templates.ObjectPage',
+                    'options': {
+                        'settings': {
+                            'entitySet': 'Booking'
+                        }
+                    }
+                },
+                ...(!testCase.isListReport && testCase.isNewPageUnavailable
+                    ? {
+                          BookSupplementObjectPage: {
+                              'id': 'BookSupplementObjectPage',
+                              'name': 'sap.fe.templates.ObjectPage',
+                              'options': {
+                                  'settings': testCase.isContextPathDefined
+                                      ? {
+                                            'contextPath': '/Travel/_Booking/_BookSupplement'
+                                        }
+                                      : {
+                                            'entitySet': 'BookingSupplement'
+                                        }
+                              }
+                          }
+                      }
+                    : {})
+            };
+
+            const routes = [
+                {
+                    'pattern': ':?query:',
+                    'name': 'TravelList',
+                    'target': 'TravelList'
+                },
+                ...(testCase.isListReport || testCase.isNewPageUnavailable
+                    ? [
+                          {
+                              'pattern': '/Travel({key}):?query:',
+                              'name': 'TravelObjectPage',
+                              'target': 'TravelObjectPage'
+                          }
+                      ]
+                    : []),
+                {
+                    'pattern': '/Travel({key})/_Booking({key1}):?query:',
+                    'name': testCase.isNoRouteFound ? 'unknown' : 'BookingObjectPage',
+                    'target': 'BookingObjectPage'
+                }
+            ];
+            jest.spyOn(rtaMock.getRootControlInstance(), 'getManifest').mockReturnValue({
+                'sap.ui5': {
+                    routing: { routes, targets }
+                }
+            });
+            jest.spyOn(rtaMock, 'getFlexSettings').mockImplementation(() => {
+                return {
+                    projectId: 'dummyProjectId'
+                } as FlexSettings;
+            });
+
+            const dummyAppComponent = {} as unknown as AppComponentV4;
+            jest.spyOn(utils, 'getV4AppComponent').mockReturnValue(dummyAppComponent);
+
+            const metaModelMock = {
+                requestObject: jest.fn().mockImplementation((path: string) => {
+                    if (path.split('/').length > 2) {
+                        switch (path) {
+                            case '/TravelType/_Booking':
+                                return {
+                                    $isCollection: true
+                                };
+                            case '/BookingType/_BookSupplement':
+                                return {
+                                    $isCollection: true
+                                };
+                            default:
+                                return {
+                                    $isCollection: false
+                                };
+                        }
+                    } else {
+                        switch (path) {
+                            case '/Travel':
+                                return {
+                                    $Type: 'TravelType',
+                                    $NavigationPropertyBinding: {
+                                        _Booking: 'Booking',
+                                        _Agency: 'Agency'
+                                    }
+                                };
+                            case '/Booking':
+                                return {
+                                    $Type: 'BookingType',
+                                    $NavigationPropertyBinding: {
+                                        _BookSupplement: 'BookingSupplement',
+                                        _Travel: 'Travel'
+                                    }
+                                };
+                            case '/BookingSupplement':
+                                return {
+                                    $Type: 'BookingSupplementType',
+                                    $NavigationPropertyBinding: {}
+                                };
+                        }
+                    }
+                })
+            };
+            jest.spyOn(rtaMock.getRootControlInstance(), 'getModel').mockReturnValue({
+                getMetaModel: () => metaModelMock
+            } as unknown as ODataModelV4);
+
+            const registry = new FEV4QuickActionRegistry();
+            const service = new QuickActionService(
+                rtaMock,
+                new OutlineService(rtaMock, mockChangeService),
+                [registry],
+                { onStackChange: jest.fn(), getConfigurationPropertyValue: jest.fn() } as any
+            );
+
+            CommandFactory.getCommandFor.mockImplementation((control, type, value, _, settings) => {
+                return { type, value, settings };
+            });
+
+            await service.init(sendActionMock, subscribeMock);
+
+            await service.reloadQuickActions({
+                'sap.uxap.ObjectPageLayout': [
+                    {
+                        controlId: 'ObjectPage'
+                    } as any
+                ],
+                'sap.f.DynamicPage': [
+                    {
+                        controlId: 'ListReport'
+                    } as any
+                ],
+                'sap.m.NavContainer': [
+                    {
+                        controlId: 'NavContainer'
+                    } as any
+                ]
+            });
+
+            // filter out irrelevant actions
+            const actions = (sendActionMock.mock.calls[0][0].payload[0]?.actions as QuickAction[]) ?? [];
+            for (let i = actions.length - 1; i >= 0; i--) {
+                if (actions[i].title !== 'Add Subpage') {
+                    actions.splice(i, 1);
+                }
+            }
+            await subscribeMock.mock.calls[0][0](
+                executeQuickAction({
+                    id: actionId,
+                    kind: 'simple'
+                })
+            );
+
+            expect(sendActionMock).toHaveBeenNthCalledWith(
+                1,
+                quickActionListChanged([
+                    {
+                        title: testCase.isListReport ? 'LIST REPORT' : 'OBJECT PAGE',
+                        actions: !testCase.expect.toBeAvailable
+                            ? []
+                            : [
+                                  {
+                                      kind: 'simple',
+                                      id: actionId,
+                                      enabled: !!testCase.expect.toBeEnabled,
+                                      tooltip: testCase.expect.tooltip,
+                                      title: 'Add Subpage'
+                                  }
+                              ]
+                    }
+                ])
+            );
+
+            if (!testCase.expect.toBeAvailable) {
+                expect(DialogFactory.createDialog).toHaveBeenCalledTimes(0);
+            } else {
+                expect(DialogFactory.createDialog).toHaveBeenCalledWith(
+                    mockOverlay,
+                    rtaMock,
+                    'AddSubpage',
+                    undefined,
+                    {
+                        appReference: 'dummyProjectId',
+                        navProperties: testCase.isNewPageUnavailable
+                            ? []
+                            : [
+                                  testCase.isListReport
+                                      ? {
+                                            entitySet: 'Travel',
+                                            navProperty: 'Travel'
+                                        }
+                                      : {
+                                            entitySet: 'BookingSupplement',
+                                            navProperty: '_BookSupplement'
+                                        }
+                              ],
+                        pageDescriptor: {
+                            appType: 'fe-v4',
+                            appComponent: dummyAppComponent,
+                            pageId: testCase.isListReport ? 'TravelList' : 'BookingObjectPage',
+                            routePattern: testCase.isListReport ? ':?query:' : '/Travel({key})/_Booking({key1}):?query:'
+                        },
+                        title: 'ADD_SUB_PAGE_DIALOG_TITLE'
+                    },
+                    {
+                        actionName: 'add-new-subpage',
+                        telemetryEventIdentifier
+                    }
+                );
+            }
         });
     });
 });